+++ zetaclient/orchestrator/v2_bootstrap.go
@@ -2,16 +2,25 @@ package orchestrator
 
 import (
 	"context"
+	"fmt"
 
+	ethcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/onrik/ethrpc"
 	"github.com/pkg/errors"
 
+	"github.com/zeta-chain/node/pkg/chains"
 	"github.com/zeta-chain/node/zetaclient/chains/base"
 	"github.com/zeta-chain/node/zetaclient/chains/bitcoin"
 	"github.com/zeta-chain/node/zetaclient/chains/bitcoin/client"
 	btcobserver "github.com/zeta-chain/node/zetaclient/chains/bitcoin/observer"
 	btcsigner "github.com/zeta-chain/node/zetaclient/chains/bitcoin/signer"
+	"github.com/zeta-chain/node/zetaclient/chains/evm"
+	evmclient "github.com/zeta-chain/node/zetaclient/chains/evm/client"
+	evmobserver "github.com/zeta-chain/node/zetaclient/chains/evm/observer"
+	evmsigner "github.com/zeta-chain/node/zetaclient/chains/evm/signer"
 	zctx "github.com/zeta-chain/node/zetaclient/context"
 	"github.com/zeta-chain/node/zetaclient/db"
+	"github.com/zeta-chain/node/zetaclient/metrics"
 )
 
 const btcBlocksPerDay = 144
@@ -58,6 +67,64 @@ func (oc *V2) bootstrapBitcoin(ctx context.Context, chain zctx.Chain) (*bitcoin.
 	return bitcoin.New(oc.scheduler, observer, signer), nil
 }
 
+func (oc *V2) bootstrapEVM(ctx context.Context, chain zctx.Chain) (*evm.EVM, error) {
+	// should not happen
+	if !chain.IsEVM() {
+		return nil, errors.New("chain is not EVM")
+	}
+
+	app, err := zctx.FromContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	cfg, found := app.Config().GetEVMConfig(chain.ID())
+	if !found || cfg.Empty() {
+		return nil, errors.Wrap(errSkipChain, "unable to find evm config")
+	}
+
+	httpClient, err := metrics.GetInstrumentedHTTPClient(cfg.Endpoint)
+	if err != nil {
+		return nil, errors.Wrapf(err, "unable to create http client (%s)", cfg.Endpoint)
+	}
+
+	evmClient, err := evmclient.NewFromEndpoint(ctx, cfg.Endpoint)
+	if err != nil {
+		return nil, errors.Wrapf(err, "unable to create evm client (%s)", cfg.Endpoint)
+	}
+
+	evmJSONRPCClient := ethrpc.NewEthRPC(cfg.Endpoint, ethrpc.WithHttpClient(httpClient))
+
+	baseObserver, err := oc.newBaseObserver(chain, chain.Name())
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to create base observer")
+	}
+
+	observer, err := evmobserver.New(baseObserver, evmClient, evmJSONRPCClient)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to create observer")
+	}
+
+	var (
+		zetaConnectorAddress = ethcommon.HexToAddress(chain.Params().ConnectorContractAddress)
+		erc20CustodyAddress  = ethcommon.HexToAddress(chain.Params().Erc20CustodyContractAddress)
+		gatewayAddress       = ethcommon.HexToAddress(chain.Params().GatewayAddress)
+	)
+
+	signer, err := evmsigner.New(
+		oc.newBaseSigner(chain),
+		evmClient,
+		zetaConnectorAddress,
+		erc20CustodyAddress,
+		gatewayAddress,
+	)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to create signer")
+	}
+
+	return evm.New(oc.scheduler, observer, signer), nil
+}
+
 func (oc *V2) newBaseObserver(chain zctx.Chain, dbName string) (*base.Observer, error) {
 	var (
 		rawChain       = chain.RawChain()
@@ -88,4 +155,18 @@ func (oc *V2) newBaseObserver(chain zctx.Chain, dbName string) (*base.Observer, 
 
 func (oc *V2) newBaseSigner(chain zctx.Chain) *base.Signer {
 	return base.NewSigner(*chain.RawChain(), oc.deps.TSS, oc.logger.base)
+}
+
+func btcDatabaseFileName(chain chains.Chain) string {
+	// legacyBTCDatabaseFilename is the Bitcoin database file name now used in mainnet and testnet3
+	// so we keep using it here for backward compatibility
+	const legacyBTCDatabaseFilename = "btc_chain_client"
+
+	// For additional bitcoin networks, we use the chain name as the database file name
+	switch chain.ChainId {
+	case chains.BitcoinMainnet.ChainId, chains.BitcoinTestnet.ChainId:
+		return legacyBTCDatabaseFilename
+	default:
+		return fmt.Sprintf("%s_%s", legacyBTCDatabaseFilename, chain.Name)
+	}
 }